requirements.txt
streamlit
pillow

openai
python-dotenv